+++ lib/redmine_checklist.rb
@@ -1,3 +1,4 @@
 require_relative 'redmine_checklist/hooks/asset_hooks'
 require_relative 'redmine_checklist/hooks/view_hooks'
 require_relative 'redmine_checklist/patches/issue_patch'
+require_relative 'redmine_checklist/patches/issues_helper_patch'